Convert back to int a couple of high-level variables

+++ cmd/plotter/main.go
@@ -13,10 +13,10 @@ import (
 )
 
 var (
-	k        = flag.Uint64("k", 15, "Storage parameter")
+	k        = flag.Int("k", 15, "Storage parameter")
 	plotPath = flag.String("f", "plot.dat", "Final path to the plot")
 	keyPath  = flag.String("key", "", "Path to key to be used as a plot seed")
-	availMem = flag.Uint64("m", 20*1024*1024, "Max memory to use when plotting. Defaults to all OS available memory when set to zero.")
+	availMem = flag.Int("m", 20*1024*1024, "Max memory to use when plotting. Defaults to all OS available memory when set to zero.")
 )
 
 func main() {
